feat: allow plugins to handle applications that use other tools
In order to support advanced use cases like fetching helm values from a
configmap without actually adding support for that to the core of
ArgoCD, we can instead enhance the plugin API to allow plugins to handle
sync requests for Helm/Kustomize/etc.. applications.

By doing so, a plugin implementor can add RBAC to the repo server and
write their own plugins in order to combine the passed values/value
files with ones in a configmap or secret, or they can do something like
make a network requests to a separate microservice that has that
responsibility.  The security of those requests could is outside of the
scope of this, but could be handled via mTLS/NetworkPolicy, and the
authenticity of a request for values could be handled via signing the
application and then verifying the provenance of an application.